update MNIST mean + std to float

+++ android/app/src/main/java/io/thomasduffy/torchmnist/TensorUtils.kt
@@ -1,17 +1,18 @@
 package io.thomasduffy.torchmnist
 
+import android.util.Log
 import org.pytorch.Tensor
 
 class TensorUtils {
     companion object Converter {
-        val MNIST_MEAN = 0.3081
-        val MNIST_STD = 0.1307
+        val MNIST_MEAN = 0.3081f
+        val MNIST_STD = 0.1307f
         val BLANK = - MNIST_STD / MNIST_MEAN
         val FILLED = (1.0f - MNIST_STD) / MNIST_MEAN
         val IMG_SIZE = 28
 
         fun convertToTensor(points: MutableList<MutableList<Pair<Float, Float>>>, h: Int, w: Int): Tensor {
-            val inputs = Array<Double>(IMG_SIZE * IMG_SIZE) {_ -> BLANK}
+            val inputs = Array<Float>(IMG_SIZE * IMG_SIZE) {_ -> BLANK}
             points.forEach {segment ->
                 segment.forEach {pair ->
                     val pX = pair.first.toInt()
